Automatically abstract interfaces required by annotated lambda binders
With this change, we automatically elaborate a type such as
```
(n : Type) ?-> (a : Type) ?-> Table n a -> a
```
into
```
(n : Type) ?-> (a : Type) ?-> Ix n ?=> Table n a -> a
```
(assuming `def Table (n : Type) (_ : Ix n) ?=> (a : Type) : Type = n=>a`)

In the near future this should let us start enforcing the index set
constraints in a backwards-compatible and non-verbose way, since any
type appearing to the left of `=>` will get constrained automatically.

This will be convenient for associated types too, because e.g. mentioning
`TangentSpace a` in a type will automatically add a constraint `Manifold a`
to the type signature.