Fix merge.
Required some changes to pi binders to mimic changes from main.

I disabled the safe names path for now because it's getting hard to sustain both
options. This will be enough to merge with main. Then we can turn the flag on in
a separate PR once we've fixed the remaining bugs.