Standardize on using Fallible everywhere instead of MonadFail in some places.
The reason is that we want good errors messages from `Builder` when we build a
block whose type leaks local vars. So we have to give it a `Fallible` instance,
not just a `MonadFail` one.

But it also means giving `Fallible` instances to `[]`/`Maybe` that just swallow
the error messages, which is a bit strange. On balance I think it's fine, but
maybe we'll eventually find a better solution.