Rewrite cheap reduction pass used for normalizing types.
The most recent version was very aggressive. Starting with an atom or
expression, it inlined and recursively reduced anything relevant in scope. This
turned out to be far too expensive. The linalg tests in particular took a really
long time (so long that I never saw them finish), because it kept traversing big
functions like `lu` in an attempt to reduce integer-valued expressions as much
as possible. I noticed because it ran very quickly when I stubbed out the
definition of `lu` with `todo`!

This change goes back to a much more conservative strategy. There's now a "push"
component, where we reduce blocks by reducing each decl, top-to-bottom, adding
the results to the substitution. Then there's a "pull" component where we look
at the definitions of variables we're trying to reduce, like `Fin`. The push
component tries hard to reduce eveything, including applications but the pull
component only inlines atoms.

Some programs now fail to type check. I had to tweak a function in sort.dx, and
I commented out the "automatic quantification" typeclass tests.

This is still just a temporary solution. We need to figure out exactly what our
rules should be for reducing type expressions. It's one of the bigger design
choices in a dependent type system like this.