Add duelling `Mut/Immut` constraints to the in-place scope extension system.
The `Emits` constraint for the builder monad has been working really well. This
change adds a similar mechanism to all the in-place monads.

Specifically, we introduce a pair of constraints: `Mut n`, meaning "n may be
updated in-place" and `Immut n`, meaning "n may not be updated in place". If
we have the `Immut n` constraint, we're free to access bindings and scopes
directly, and use all the "pure" machinery. We can lift an `Immut n` computation
into a `Mut n` comuptation using the rank-2 `liftImmut`, but we avoid having to
have rank-2 lifters everywhere else.

It also lets us remove the "zonkable" hacks.