Expand auto-quantification to interfaces needed by nested expressions
Presenting the previous implementation with a type such as `a=>b=>c`,
it would only auto-quantify over `Ix a`, but would fail to do so for
`Ix b`. The new implementation gathers the requirements from the whole
elaboration of the UType, for as long as the interfaces are not required
for the elaboration result (in particular they can be phantom requirements).