Start implementing safe-names Dest builder in Imp translation.
This turned out to be more involved than expected because we were previously
(knowingly) sloppy with names. When creating a Dest, we generate two sorts of
names: (1) names for pointers to buffers that we learn are needed, and (2) names
for ordinary decl emissions to compute pointer offsets inside table lambdas.
Although the emissions are interleaved, the intended scopes are speparate. So we
have to do some manual hoisting at the end of it all. Eventually it might be
good to have a system for emitting into multiple scope levels within a single
monad. This could be useful when we hoist constant subexpressions out of loops
and lambdas, for example.