Make sure data and type constructors are always fully applied
Previously we allowed them to be partially applied, in which case they
would have function types. This approach is a little surprising and
makes the implementation of interface dependencies for data definitions
more difficult, so this patch makes it so that data and type
constructors are always elaborated as nested lambda expressions that
contain fully applied `DataCon`s or `TypeCon`s in their bodies.

Note that this doesn't really change anything important. It's only a
setup for follow-up work.