Rename "Env" -> "Subst", "Bindings" -> "Env", and "inject" -> "sink"
We'd been planning these changes for some time and this seems like a good time
to do it.

"Bindings" was originally just a mapping from names to their definitions. But
it's grown to include other parts of a general static environment: synthesis
candidates, allowed effects and the source map. Meanwhile, what we used to call
"Env", a mapping from names in one name space to values in another, turns out to
be exclusively used for substitutions.

"sink" is a better name than "inject" because it's more evocative, it's less
overloaded, it's dual to "hoist", and it's shorter.