Safe-names translation rule for runtime error.
I did it a bit differently this time. `ThrowError` just translates to
`IThrowError` and we return a nonsense atom loaded from an uninitialized dest.

In the unsafe-names version, we put the whole Imp translation under an `ExceptT`
transformer, and we translated `ThrowError` by throwing an error in the
translation. At various places we'd catch that error and emit `IThrowError`. It
was a bit tricky to know when it was necessary to handle errors (we had comments
like "XXX: This does not handle errors that happen inside the block!"). We made
that change in 259e84d9b0 to get rid of the special `AnyVal` value. In the
current change, we're basically just simulating `AnyVal` with a read from an
uninitialized dest which happens to occur on a dead control flow path. I think
it's fine, but I might be missing something.