Allow substituting B-kinded things by themselves.
I'm trying to make it easier to go under a binder when you need to. For example,
imagine you have an `Abs b e1 n` and an `e2 n`. And you want to sink the `e2`
under the binder to make an `Abs b (PairE e1 e2) n`. The `DepPairTy` case of
`makeDestRec` shows how hard this is currently if you're not already in a
substitution monad.

With this change, `refreshBinders` gives you a fragment of a subsitution you can
use as you please. You can extend a monadic substitution if you have one, but
you can also just apply it directly.