Handle the full set of built-in index set types
We implement the same three functions (index set size + to/from int) in both
Builder and Type (for literals) which is a lot of duplication. At the cost of
some import cycles we could handle the literal versions by running the
interpreter over the builder versions. But I'm hoping the problem will go away
once we make index sets a proper user-level type class, and then we'll be able
to remove both versions.