Remove `ProjectElt` atoms from index type nests before converting to polynomials.
The logic in Algebra can't handle ProjectElt, which was causing programs like
this to fail: `concat for i:(Fin 4). AsList _ [(Foo2, Foo2)]`. We now do a pass
where we emit decls for those ProjectElt atoms so we just have names instead.
I'd really like to remove ProjectElt, or at least replace it with a much more
specialized version, so we can avoid hacks like these.