Split the `Builder` class into `Builder` and `ScopedBuilder`.
Sometimes we can implement `emitDecl` but not `buildScoped` (`DestM` is one
example). This change puts the two methods in separate classes so we can have
one without the other. Previously I just left `buildScoped` as `undefined` for
`DestM` but we were actually hitting it and this was the easiest way for me to
figure out where I was going wrong.