Add support for interface constraints on `data` types
This should be the last blocker for rolling out the `Ix` constraint in
surface language. Without this syntax it would be impossible to define
a type such as:
```
data MyData n a = MkMyData (n=>a)
```
because there would be no `Ix n` constraint to satisfy the left hand
side of `=>`. Instead, this type should be defined as:
```
data MyData n a [Ix n] = MkMyData (n=>a)
```