Allow local type aliases in both surface language and core
Turns out that at some point we might have lost (or perhaps never
acquired?) the ability to have local type aliases due to limitations
of our type equality relation and cheap reduction. We still treat alpha
equivalence as the first check, but in case it fails we do fall back to
cheap normalization before we conclude the check as unsuccessful.