Add alpha-renaming-invariant hashing.
An expression in some scope `n` can contain both free names, of type `Name c n`,
and alpha-renamable bound names, whose binders are in the expression itself. We
hash the free names according to the actual name string but we use a sort of de
Bruijn scheme for the bound names. Occurrences of these names are hashed
according to the lexical position of their binder, numbered from outside to
inside (the opposite of the de Bruijn convention).

This is equivalent to the "locally nameless" approach in the taxonomy laid out
in "Hashing Modulo Alpha-Equivalence" by Krzysztof Maziarz and co (section 2.5).
As that paper describes, it has the downside that the hash of `\x. e` is not
computed in terms of the hash of `e`, so if you need both hashes you need to do
two separate passes.

I don't know if that's going to be a problem or not, but I'm planning to use the
hashes for some CSE stuff soon and we might find out then.