Add support for user-defined index sets
... and make all existing index sets defined in user code. No methods
are implemented inside the compiler anymore! This is great in that it
opens up lots of possibilities (a very simple one is that the `(|)` type
from prelude can finally be an index set), but it does come with a
pretty significant slowdown in the compiler (likely due to lack of
caching).