Don't use Algebra for rectangular table types
The Algebra module is neat, but it's not fully general. In the simple
case of rectangular (i.e. non-dependent) index nests the arithmetic is
simple enough that we don't need to engage it at all. In particular this
should let us the arbitrarily complex Ix implementations that users
will soon be able to throw at the compiler without crasing due to
partiality of our simplifier.