Remove the `Module` abstraction. Instead, work directly with top-level decls and blocks.
The `Module` abstraction originally served two purposes: (1) it let us structure
the compiler as a fixed sequence of non-monadic module-to-module passes and (2)
it provided local name spaces so we didn't have to worry about interference
between top-level names arising from different source blocks.

Our naming rewrite completely solves (2). Hooray! That leaves (1), which I've
come to see as a non-goal. As we add things like caching, it actually helps to
allow the passes to modify state (e.g. adding cache entries). And as long as we
do it with our usual read/extend pattern, it doesn't get in the way of parallel
compilation or persistent compilation caching or anything like that.

The new top-level structure processes top-level decls one by one (as before,
just no longer wrapped as a "module"), but quickly isolates a `Block n`
representing the part of the decl that actually requires compilation and
compute. When we come across a compute-free decls like a data definition, we
handle it swiftly and move on. With ordinary decls, type inference turns the rhs
into a core Block, which is the only sort of thing that the downstream passes
have to contend with. This is an simpler object to work with than the unwieldy
B-kinded `Nest Decl n l` that we had before.

One awkwardness is that we still have to handle top-level pattern-matching (like
`(x, y) = foo 10`). Previously we generated code for that. In the new model, we
only generate code for the right hand side, `foo 10`, and then we do the
pattern-matching logic in an interpreter once we've executed that code. This is
guaranteed to succeed because it's after type inference.