Log emissions made during passes.
Now that passes emit bindings directly rather than bundling them into a module,
it's important to track them as they happen. For example, without this, logging
the result of a data definition just gives the name substitution and nothing
else.

As part of this change I ended up making the checking a bit more uniform, since
we usually do logging and checking together. It required defining dedicated
types for the result of Imp/Simp so that they could be given checkable
instances. Probably not a bad thing.