Add an extra layer of Ix caching in Imp and thunk the size method
Those two seemed to cause the largest slowdowns, first because it
required repeated simplification of the same snippet of code and the
second because it ended up spewing lots of dead decls any time we
syntheiszed an Ix dictionary. It's not the most elegant, but this is the
best we can do right now.