Define unzipping and transposition rules for App in the prototype.
Unzipping in the presence of App is a bit subtle, because we need to
typecheck the unzipped function body in order to know what type to
declare for the tape.  But the Program in which we need to typecheck
it must have all callees already unzipped!  For now, I tied the knot
with Haskell's laziness (instead of trying to explicitly compute the
call graph); seems to work on the examples present in the test suite.

I also carry around the original (pre-unzipping) program to look up
the number of results callees return, so that I know how many
variables to allocate to hold them.

Finally, added better error checking in case an unzipped program fails
to typecheck (for debugging the unzip transformation).