Compile standalone monomorphic first order functions based on @noinline decorator.
We had a version of this working before but it would still recompile the
standalone function from scratch in every top level block that used it. This
time, we're truly compiling just once. We generate an object file which gets
linked as needed. These object files get saved to disk in the cache along with
everything else, so we shouldn't need to recompile every time we start a new
process.