Make the set of name colors a bit more open.
Previously, adding a new name color meant adding a case to `NameColorRep`,
`DynamicColor` and so on. There were 9 different places in Name.hs alone.

With this change, we use the term-level version of `C` itself as the term-level
tag, replacing `NameColorRep`. And we avoid `DynamicColor` by implementing the
`Store` instance for `WithColor` directly rather than via generics. There's
still an exhaustive enumeration in `interpretColor` but this can't be avoided if
we want to be able to deserialize. Otherwise it should now be less work to add
new name colors.

One cost is that there are a lot more `NameColor c` constraints, so I renamed
`NameColor` to `Name` to save a few characters.