Add limited support for first-class record labels
This change extends our record system with an option to have a single
field with a statically unknown label in front of the static labels and
the row extension (which we supported previously). For example the type
```
{@v: Int & v: Float & ...rest}
```
represents to a record with a field of type `Int` and a _dynamic label name_,
specified by the variable `v` (which should be of type `Label` --- that's
another addition). The record also has a field with statically known
name called `v` (for maximum confusion) of type `Float` and a
polymorphic tail of other fields given by `rest`.

`Label`-typed values can be created using the `##name` syntax, similarly
to how we use `#name` for lenses. We might want to revisit this sugar in
the future, but this was the path of least resistance for now.

An interesting property of our type system is that it allows both
dynamic labels and label shaddowing, which to the best of my knowledge
is unique. Doing so lets us avoid dealing with lots of awkward typeclass
constraints of the form "dynamic label disjoint form record". It's
possible that there are some contradictions lurking there, but so far it
does seem to be sound. The trick was to forbid the projection of static
fields that are preceded by a dynamically named one. This means that to
unpack a record type as shown above, one _always_ has to unpack the `@v`
field before they project out the `v` field. So for example this is a
type error:
```
{v=vf, ...} = record
```
but this is allowed
```
{@v=dvf, v=vf, ...} = record
```
This makes it possible to resolve the unpacks statically without having
to worry whether the variable `v` could later get instantiaed to `##v`.