Checking in a WIP because I'm about to set aside this branch for a bit.
I got stuck trying to rework the InplaceT API. I had something like this for
producing a scoped nest of decls

```
scopedInplace
  :: SinkableE e
  => (forall l. (Mut l, DExt n l) => m l (e l))
  -> (forall l.         DExt n l => decls n l -> e l -> m l a)
  -> m n a
```

But if you're then going to re-emit some of those decls (as we do in inference)
then you need `DistinctAbs decls e n`, which we can't have because it would need
`Immut`.

But the only reason we like to emit `DistinctAbs` instead of `Abs` is to avoid
the cost of renaming everything. Maybe if we just had a fast path for renamings
that (dynamically) happen to be no-ops, then we wouldn't have to worry about
it...