Remove the distinctness-carrying `DistinctAbs` in favor or `Abs`.
You can always obtain a `DistinctAbs` from an `Abs` by doing a name-freshening
pass. But this can be expensive and we want to avoid doing it unnecessarily.
That's why many of our APIs (especially InplaceT ones) accepted or returned
`DistinctAbs`. But `DistinctAbs` is an awkward object to work with because it's
not sinkable -- sinking it would invalidate the distinctness. In particular,
this means you can never have access to a `DistinctAbs b e n` if you're in an
in-place monad with `Mut n`. We worked around this with complicated APIs, like
this:

```
extendInplaceT
  :: forall m b d e n.
     (ExtOutMap b d, Monad m, SinkableE e)
  => Mut n
  => (forall l. (Immut l, DExt n l) => InplaceT b d m l (DistinctAbs d e l))
  -> InplaceT b d m n (e n)
```

But there's another way! We can have a short-circuiting version of `refreshAbs`
that checks dynamically whether the binders happen to be fresh already, and just
applies `unsafeCoerceE` to the body in that case. The static type system may
force you to refresh an abs that happens to already be fresh. But the redundant
refreshing pass will be cheap.

A downside is that performance now relies on a dynamic property rather than a
static one. (Though correctness is still guaranteed statically.) But it's not
very hard to reason about. And it might even yield wins in other cases, where
refreshment is unnecessary because names happen to be sufficiently fresh
already. We can even imagine trying to deliberately make names weirder than
necessary to trigger the fast path more often. It becomes like a hash table with
a fallback path for hash collisions. The fallback path is important for
correctness, but rarely exercised.

This commit removes `DistinctAbs` but doesn't implement the short-circuit
optimization. That's coming up next.