Fix a performance bug in handling scopes.
We've had this one for a while but I only just noticed it. In most of the
EnvReader monads, every time we generate a fresh name we compute the current
scope from the current set of bindings. This was supposed to be an O(1)
operation but actually we were applying `fmap` to the Map backing the env, which
is O(N)! This fixes that bug by storing the scope separately from the env.
Prelude eval is now slightly faster than it was before we made the
dynamically-distinct-shortcut change.