Fix a bug in dynamic distinctness check.
We just added an optimization to `refreshAbsPure` that dynamically checks
whether the binders are already fresh wrt the scope and does nothing in that
case. But the check wasn't strict enough. We checked that the names in the
binders were distinct with respect to the scope, but we didn't check that they
were distinct with respect to *each other*.

To do the check correctly, we need envs and scopes to carry information about
shadowed names. Roughly, we make envs and scopes behave like `Map Name [a]`
instead of `Map Name a`.