Add prelude-importing code in parser.
We had a lot of logic for managing when to import the prelude and when not to
(e.g. we don't import the prelude when we're evaluating the prelude!). In my
last commit I even missed a spot: I didn't consider the prelude when considering
module dependencies. I think it's easiest if we just do it in one place, which
may as well be the parser. (Almost one place. We also need to handle it
explicitly in the repl, because there we parse things line by line.)