The matrix2x2 example discovered a bug in splitTangents.
The bug was that, if the first expression had free variables in common
with the remainder, splitTangents would drop the variables that the
first expression referred to alone.  Now fixed.

Also tweak the `gradient` test helper to accept Linear A values rather
than Haskell values so it can applied effectively at more Linear A
types.