Fix a bug to do with interface source names leaking internal names.
We synthesize a type for dictionaries corresponding to interfaces. These include
a source name for printing (it also ends up being used in equality checking due
to deriving Eq). But we had been obtaining this source name as the pretty-print
of an internal name. Loading some cached bindings meant we ended up with
different internal names (i.e. `Ix2` instead of `Ix` because `Ix` was already
taken) which broke things. The fix is to just to use original source name from
the source program rather than guess it by printing the internal name.