Tweak a few quine tests.
The problem is that they expose internal names which can now be different
depending on what's been cached. We shouldn't be exposing these names in tests.
But as a workaround, if we make the string component of the names weirder, we
can get away with it for a bit longer.