Read the prelude from the file system rather than baking it into the compiler.
I'm doing this now because the macOS CI is failing due to a cyclic dependency
between Resources and the main dex library. But I think it's better anway.
Reading the prelude from the file system means you can edit it without
recompiling or without even restarting the dex process in the notebook view.