Try representing raw names using `Int` instead of `(String, Int)`.
The thought was that this would let us use `IntMap`, which is said to be much
faster than `Map`. We can still have human-readable names with an encoding
shceme like the following. We use 7-bytes of the 64-bit in for ascii characters,
leaving the other byte for the counter, falling back to non-pretty counter-based
names once we exhaust the 256 options, with the first bit reserved to indicate
which representation we're using. This commit doesn't try to use that encoding.
It was just to see how big the win would be. And the result was...not very big!
Evaluating the prelude is about as fast as before. So I'm putting this aside for
now.