Perform algebraic simplification in Simplify
Computing the size of `Fin n` for a known constant `n` takes three
instructions that can be constant folded. It used to happen when
`indexSetSize` was implemnted in the compiler using the builder methods,
but it's not happening for the current implementation in user-space!

The simple redirection of simplified ops through relevant builder
methods closes the gap again (and interestingly enough speeds up our
compilation quite a bit in certain scenarios).