Add a reverse nest (snoc nest) for linear-time decl accumulation
Back when we used the classic implementation of Writer for InplaceT this
didn't make a big difference, but now that the extensions are generally
always on the right, it does! With this change GC copies 12% fewer bytes
over a single run of the kernel regression example, leading to a ~5%
end-to-end speedup!