Simplify the actor system to the bare essentials.
The bare essential idea of actors is that each thread is a little
server: an actor is a thread that receives information only as
messages on a particular channel, and no other thread may read from
that channel.

All the machinery involving links and ErrMsg and so forth was just
about error handling, and wasn't being used, so delete it (at least
for now).

Also eliminated the Actor monad, since it's simpler for each actor to
just close over its input channel.