Distinguish tables and functions with separate constructors rather than arrows.
We end up treating them quite differently anyway so it's a reasonable thing to
do. The reason to do it now is that I want to put the `Ix` dictionary in the
table type itself rather than have it be a phantom constraint and re-synthesize
it when we need it.

It also makes the (Haskell) types a bit more precise because we syntactically
rule out table types with effects.