Scan standard library, benchmarks, and examples to reduce hanging indentation.
Hanging indentation is when the amount a line is indented depends on a
name, usually one that occurs on the previous line.  For example,

```
def some_long_function (arg1: type1)
                       (arg2: type2)
                       : return =
  body
```

It's not clear that this actually aids readability much, relative to
just breaking the line after `some_long_function`; and it's highly
fragile to automatic renames that change the length of the function
name.  Ergo, it seemed best to just not do this in our code.

The specific case of function signatures with long argument lists is
fairly common, and merits some comment.  Following a few examples of
Dougal's, I adopted a convention of indenting the argument list six
spaces, and aligning the : return on a line by itself.
Conventionally, I also leave implicit arguments on the same line as
the function name, if they fit.  This ends up looking like

```
def some_long_function
      (arg1: type1)
      (arg2: type2)
      : return =
  body
```

The signature indentation is not 2 spaces, so it doesn't align with
the beginning of the body.  It's also not 4 spaces, so it doesn't
align with the function name (since the `def` keyword is 3 characters
long).