Add a fast-path for retrieval of Dex strings
Many of our examples produce large Dex strings that are to be
interpreted by the top level runtime as HTML to be displayed. However,
as profiling indicates, retrieving those strings is pretty much the
single biggest overhead, taking almost a half of all run-time on the
kernel regression example.

In this commit, I added a fast-path that tries to pattern match on the
atom that will (always?) be produced by running the Dex program. If it
succeeds, we skip the expensive index generation and table body
interpretation. Instead, we can simply retrieve the data with a builtin
Haskell function. With this patch the overhead of this function becomes
neglibile!