Get superclasses from the given dict's type rather than trying all projections.
We were doing a very silly thing in dictionary synthesis. When we first
implemented type classes we wrote superclass projections by hand. To obtain the
superclasses for a particular given, we would try to apply every superclass
projection available, almost all of which would fail to type check. Maybe that
made sense at the time, but it's completely ridiculous now. A dictionary's
superclasses are now plainly described in the dictionary's type.

This speeds up prelude evaluation by 25%.

I had to special-case the `AccumMonoid` dictionary because we still define that
one by hand, so its representation isn't like the others.