Use a CPS-style writer for dest accumulation in InplaceT
Writer-style accumulation turns out to be really annoying. It produces
weird associations of mconcat and any trivial monadic action causes the
bind to extend with mempty (which is obviously a no-op). This changes
InplaceT --- one of our most fundamental monads --- to use a CPS-style:
instead of being a writer in decls (`env -> (a, decls, env)`) it is now a
stateful operation in decls (`env -> decls -> (a, decls, env)`).

This does not have a dramatic effect on the run-time (~3%?), but it does
reduce the number of total allocations by 10%.