Use a compact encoding of strings into Names
Now, this makes quite a difference! A 10% end to end speedup! 10% fewer
overall allocations! _25% fewer bytes copied by GC_! It wasn't for free,
since we do loose the ability to represent longer strings in names and
the implementation requires quite a bit of bit manupulation, but seems
to be worth it!

Note that this is based on the initial implementation and ideas of
dougalm@.