Fix #833 with Dougal.
The root of the problem is that Dex used to capture lambda-bound
variables during defunctionalization; but that's actually too
conservative, because the argument (unlike anything from the enclosing
environment) will actually be available when the lambda is
reconstructed.

This change avoids doing that, so the return type of the lambda need
no longer be hoisted past its binder, and thereby doesn't trigger the
observed leak error.

Co-authored-by: @dougalm <dougalm@google.com>