Fix bug 841, linearization of triangular table.
Defunctionalization is supposed to only capture local variables.  We
used to also be capturing all variables in the types of those local
variables, but we should only capture _local_ variables in the types
of local variables.

Not sure why exactly this come up during linearization of a function
with a local triangular table (such as the LU decomposition), but it
did.

Co-authored-by: @dougalm <dougalm@google.com>