Make EitherE6 faster
We use it all over the place in GenericE instances, but instead of being
a flat ADT it's actually a nest of binary sum types! To inject the last
case into the generic representation one has to allocate 6 extra
constructors, and that obviously adds up. Especially given that GHC is
currently unwilling to optimize out toE/fromE. And even when I try hard
to get it to simplify those away, GHC becomes very shy about inlining
things into very deep case statements.

So, overall it's a win all accross the board. If we simplify away
GenericE our code will be faster. And if we don't... it will still be
faster! With this change, we allocate On the kernel regression example we
allocate 15% less data overall, letting us save on 15% of gen-0 garbage
collections. It doesn't seem to translate to huge performance
improvements, but it does seem to let us get a percent or two
end-to-end.