Lower LLVM optimization level
I put together the previous (very expensive!) optimization pipeline
while trying to get LLVM to vectorize some of the code we emit. But
that's not a good idea. We should just emit vector code ourselves,
instead of relying on LLVM to do a good job for us. The pipeline was so
expensive that it ended up taking even _half_ of the entire run-time on
the kernel regression example, completely overwhelming the actual cost
of execution.

This lowers the optimization level to `-O1`. It might lower the run-time
performance of Dex programs as they are, but compile-time overheads seem
to be the major issue right now. In any case, we should add a family of
`-O` flags to Dex as well, to allow users to select the optimization
level for themselves.